Ensure that the scroll view decelerates vs. bounces when scrolling along a single axis
If the scroll view is exactly as tall as its content along one axis, then the
content offset along that axis will be 0 as will the minimum and maximum
edges of the bounds along that axis, and
`contentOffsetBounds.contains(contentOffset)` will fail, because
it considers coordinates lying on the maximum edges to be outside the rect.
Enlarge the bounds along the problematic axis to avoid this.