Don’t decelerate the scroll view if the user had stopped dragging
Previously, the scroll view would always do a deceleration animation when the
user finished dragging, even if the user had stopped dragging and held at a
point; whereas UIScrollView will simply pause at that point in such a case.

This problem is because pan gesture recognizers always report a non-zero
terminal velocity: https://stackoverflow.com/questions/19092375/how-to-determine-true-end-velocity-of-pan-gesture.
To fix this, we now discard the terminal velocity if it has been sufficiently long
since the user last dragged.

Passing `.zero` to `completeGesture(withVelocity:)` does not affect whether
the scroll view bounces when the gesture ends—it still will, if it needs to in
order to return the content offset to within the scroll view’s bounds.